dvc-demo/requirements.txt
scikit-learn
pandas
PyYAML